Detach aggregate/ error module from mod.rs, use separate file
