+++ src/executor/aggregate/error.rs
@@ -1,0 +1,18 @@
+use serde::Serialize;
+use std::fmt::Debug;
+use thiserror::Error;
+
+#[derive(Error, Serialize, Debug, PartialEq)]
+pub enum AggregateError {
+    #[error("unsupported compound identifier: {0}")]
+    UnsupportedCompoundIdentifier(String),
+
+    #[error("unsupported aggregation: {0}")]
+    UnsupportedAggregation(String),
+
+    #[error("only identifier is allowed in aggregation")]
+    OnlyIdentifierAllowed,
+
+    #[error("unreachable")]
+    Unreachable,
+}
